reworked laser setpoint to use ushort, automatically apply at startup, added SiG math for laser temp

+++ WasatchNET/AndorSpectrometer.cs
@@ -405,7 +405,7 @@ namespace WasatchNET
 
         public override ushort laserTemperatureRaw { get => 0; }
 
-        public override byte laserTemperatureSetpointRaw { get => 0; }
+        public override ushort laserTemperatureSetpointRaw { get => 0; }
 
         public override float batteryPercentage
         {
