+++ WasatchNET/BoulderSpectrometer.cs
@@ -800,7 +800,7 @@ namespace WasatchNET
 
         public override ushort laserTemperatureRaw { get => 0; }
 
-        public override byte laserTemperatureSetpointRaw { get => 0; }
+        public override ushort laserTemperatureSetpointRaw { get => 0; }
 
         public override float batteryPercentage
         {
